refactor of campaigns files
