+++ src/app/features/public/public.module.ts
@@ -23,8 +23,10 @@ import { RegisterFormComponent } from './pages/register-form/register-form.compo
 import { DonationsHeaderComponent } from './pages/donations/donations-header/donations-header.component';
 import { LeafletMapComponent } from './pages/contact/components/leaflet-map/leaflet-map.component';
 import { LeafletModule } from '@asymmetrik/ngx-leaflet';
-import { CampaignsSchoolComponent } from './pages/campaigns/campaigns-school.component';
 import { ToysComponent } from './pages/campaigns/toys/toys.component';
+import { CampaignsSchoolComponent } from './pages/campaigns/back-to-school/campaigns-school.component';
+
+
 @NgModule({
   declarations: [
     PublicComponent,
